+++ src/modules/tooltip/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './tooltip/tooltip.directive';
+export * from './tooltipTemplate/tooltipTemplate.directive';