+++ src/modules/tooltip/directives/tooltipTemplate/tooltipTemplate.directive.ts
@@ -1,0 +1,16 @@
+import {Directive, TemplateRef} from '@angular/core';
+
+/**
+ * Directive used for obtaining custom tooltip template
+ */
+@Directive(
+{
+    selector: '[tooltipTemplate]'
+})
+export class TooltipTemplateDirective<TData = any>
+{
+    //######################### constructor #########################
+    constructor(public template: TemplateRef<TData>)
+    {
+    }
+}