+++ src/modules/tooltip/index.ts
@@ -1,0 +1,5 @@
+export * from './components';
+export * from './directives';
+export * from './modules';
+export * from './misc/tokens';
+export * from './misc/tooltip.interface';