+++ src/modules/tooltip/misc/tokens.ts
@@ -1,0 +1,8 @@
+import {InjectionToken} from '@angular/core';
+
+import {TooltipOptions} from './tooltip.interface';
+
+/**
+ * Injection token used for injecting tooltip options
+ */
+export const TOOLTIP_OPTIONS: InjectionToken<Partial<TooltipOptions>> = new InjectionToken<Partial<TooltipOptions>>('TOOLTIP_OPTIONS');